update makefile to include all go files and uncomment wip docker build
