+++ .drone.yml
@@ -10,29 +10,30 @@ pipeline:
       - make build
 
 
-  # Docker build will be added at a later stage 
-  # docker-latest:
-  #   image: plugins/docker
-  #   repo: registry.usw.co/cloud/hermod
-  #   tags:
-  #     - latest
-  #   when:
-  #     event: push
-  #     branch: master
+  # Docker build will be added at a later stage
+  # https://rvu.atlassian.net/browse/AIRSHIP-386
+  docker-latest:
+    image: plugins/docker
+    repo: registry.usw.co/cloud/hermod
+    tags:
+      - latest
+    when:
+      event: push
+      branch: master
 
-  # docker-tagged:
-  #   image: plugins/docker
-  #   repo: registry.usw.co/cloud/hermod
-  #   tags:
-  #     - ${DRONE_TAG}
-  #   when:
-  #     event: tag
+  docker-tagged:
+    image: plugins/docker
+    repo: registry.usw.co/cloud/hermod
+    tags:
+      - ${DRONE_TAG}
+    when:
+      event: tag
 
-  # docker-test:
-  #   image: plugins/docker
-  #   when:
-  #     event: push
-  #   repo: registry.usw.co/cloud/hermod
-  #   tags:
-  #     - testing
-  #     - ${DRONE_COMMIT_SHA}
+  docker-test:
+    image: plugins/docker
+    when:
+      event: push
+    repo: registry.usw.co/cloud/hermod
+    tags:
+      - testing
+      - ${DRONE_COMMIT_SHA}
