+++ main.go
@@ -48,17 +48,3 @@ func main() {
 
 	watcher.Run(ctx, stopCh)
 }
-
-func configureLogger(logLevel string) {
-	switch logLevel {
-	case "debug":
-		log.SetLevel(log.DebugLevel)
-	case "info":
-		log.SetLevel(log.InfoLevel)
-	case "warn":
-		log.SetLevel(log.WarnLevel)
-	case "error":
-		log.SetLevel(log.ErrorLevel)
-	}
-	log.SetFormatter(&log.JSONFormatter{})
-}
