+++ utils.go
@@ -1,0 +1,17 @@
+package main
+
+import log "github.com/sirupsen/logrus"
+
+func configureLogger(logLevel string) {
+	switch logLevel {
+	case "debug":
+		log.SetLevel(log.DebugLevel)
+	case "info":
+		log.SetLevel(log.InfoLevel)
+	case "warn":
+		log.SetLevel(log.WarnLevel)
+	case "error":
+		log.SetLevel(log.ErrorLevel)
+	}
+	log.SetFormatter(&log.JSONFormatter{})
+}
